Final?
Refactored everything so that the select icon and item bg image are seperate so everything is more contained where it should be. Also, new keybinds made to be more usable. Scroll wheel to change selection and 'e' to use an item. Number swapping still words the same. Also made button for dragging item bigger so it is easier to drag and drop. Overall inventory feels smoother.