Added Jonathan's Train obstacle
Level Designer, jonathan, wanted me to add a obstacle to for their idea of a level. Right now the area is based on mouse entered for the collision box, to change it with a player character, change connect() in Area2D to "entered_area2D" in "mouse_entered" for it to work and put the collision in the stairs where they would be when the character spawns in. and then you can change the sprite to an actual train sprite