MAJOR Stalker Art and Animation Update
Added all existing art for stalker, and added in all animations EXCEPT THE RANGED ATTACK ANIMATION, AS RANGED FUNCTIONALITY DOES NOT EXIST. The animation is implemented, but there is no call to it. Next update will be adding the Stalker's ranged attack. Current issue is that the camera is very zoomed in, and thus the Ranged/Melee and Melee/Ranged swap happen too frequently. Current fix will be that the Ranged transform happens at 90 units, but the Melee happens at 80, so it will not teeter between the two as frequently. Camera size should still be discussed further. Another issue just discovered is that the Ranged to Melee is slightly offset; thus when the Stalker swaps to melee it looks like it bounces slightly. This is most likely an issue with the different sprite sizes; can fix by manually reimporting the frames and resizing. Will do later.